+++ add-performance-global.js
@@ -1,5 +1,0 @@
-import { performance } from 'node:perf_hooks';
-
-if ( !( 'performance' in global ) ) {
-	global.performance = performance;
-}