+++ node.js
@@ -1,6 +1,5 @@
 import { AxiosSession } from './axios.js';
 import { mixCombiningSessionInto } from './combine.js';
-import './add-performance-global.js';
 
 export default class NodeSession extends AxiosSession {
 }